merging feature/authorization to main

+++ controller/blogController.js
@@ -1,29 +1,18 @@
-const response = require("../utils/blogResponse")
+const response = require("../utils/response/blogResponse")
 const {BlogService} = require("../service/blogService")
 const {MongoBlogDao} = require("../dao/blog/mongoBlogDao")
-const {MysqlDao} = require("../dao/blog/mysqlBlogDao")
+const {MysqlBlogDao} = require("../dao/blog/mysqlBlogDao")
 const express = require("express"),
  negotiate = require("express-negotiate");
 
 
-const currentDatabase = new MongoBlogDao();
-// const currentDatabase = new MysqlDao();
-const blogService = new BlogService(currentDatabase);
-
-
-exports.aliasBlogs = async (req,res,next) => {
-  req.query.limit = '3';
-  req.query.sort = '-createdAt,userName';
-  req.query.fields = 'userName,blogHeadline,blogDescription,createdAt,_id'
-  next();
-}
-
-
+const mongoBlogDao = new MongoBlogDao();
+const blogService = new BlogService(mongoBlogDao);
 
 exports.getAllBlogs = async (req, res) => {
  try {
    const blogs = await blogService.getAllBlogs(req);
-   if (!blogs) throw new Error("Blogs not found");
+   if (typeof blogs === "string") return response.errorBlogResponse(404,blogs,res);
    req.negotiate({
        "application/json": function () {  response.JSONBlogResponse(200,blogs,res)},
        "application/xml" :  function () { response.JSONBlogResponse(200,blogs,res)},
@@ -36,7 +25,7 @@ exports.getAllBlogs = async (req, res) => {
 exports.getBlog = async (req, res) => {
  try {
    const blog = await blogService.getBlog(req);
-   if(!blog) throw new Error("This Blog does not Exist")
+   if(typeof blog === "string") return response.errorBlogResponse(404,blog,res)
    req.negotiate({
     "application/json": function ()  { response.JSONBlogResponse(200,blog,res)},
     "application/xml" :  function () { response.JSONBlogResponse(200,blog,res)},
@@ -60,25 +49,35 @@ exports.createBlog = async (req, res) => {
 };
 exports.updateBlog = async (req, res) => {
  try {
-  const blog = await blogService.updateBlog(req);
+  let blog = await blogService.getBlog(req);
+  if(typeof blog === "string") return response.errorBlogResponse(404,blog,res)
+
+  if (blog.username !== req.body.username) return response.errorBlogResponse(403,"Not Have permission to Update",res);
+
+  blog = await blogService.updateBlog(req);
    req.negotiate({
     "application/json": function () {  response.JSONBlogResponse(200,blog,res)},
     "application/xml" :  function () { response.JSONBlogResponse(200,blog,res)},
     "application/default": function() { response.defaultBlogReponse(200,blog,res)}
  });
 } catch (err) {
-  response.errorBlogResponse(404,"Blog updation unsuccessful",res);
+  response.errorBlogResponse(404,err.message,res);
 }
 };
 exports.deleteBlog = async (req, res) => {
  try {
+   let blog = await blogService.getBlog(req);
+   if(typeof blog === "string") return response.errorBlogResponse(404,blog,res)
+
+   if (blog.username !== req.body.username) return response.errorBlogResponse(403,"Not Have permission to delete",res);
+
    await blogService.deleteBlog(req);
    req.negotiate({
     "application/json": function () {  response.JSONBlogResponse(204,null,res)},
     "application/xml" :  function () { response.JSONBlogResponse(204,null,res)},
-    "application/default": function() { response.defaultBlogReponse(200,blogs,res)}
+    "application/default": function() { response.defaultBlogReponse(200,null,res)}
  });
 } catch (err) {
-  response.errorBlogResponse(404,"Blog deleteion Unsuccessful",res);
+  response.errorBlogResponse(404,"Blog deletion Unsuccessful",res);
 }
 };
